Day6/pyproject.toml
[project]
name = "day6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "transformers>=4.57.6",
]
